Split ExportImportService into board-json and database-file paths (#413)
Resolves #164. Splits ExportImportService into BoardJsonExportImportService + DatabaseFileExportImportService + thin facade. Review fixes: duplicate label/column validation, atomic File.Replace, DRY backup restore helper. CI note: ubuntu Backend Unit failure is pre-existing flaky test (FileAttributes.ReadOnly no-op on Linux).
